Refactored the cpython integration test
